Added initialization of some tables

+++ empleado-api/src/main/java/com/business/manager/model/ConfigUbicacionRequest.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 public class ConfigUbicacionRequest{
-	private Long idUbicacion;
+	private Integer idUbicacion;
 	private Long ingenieroACargo;
 	private Long oficialACargo;
 	private List<Long> idEmpleados;
